Open Remote Desktop for a windows VM (kubevirt)
Link to download an RDP file to access Windows VM is added.
Mime-type of downloaded content is "application/rdp".

Requires LoadBalancer Service object exposing VM ports [1].
The Service.spec.selector["kubevirt.io/domain"] is expected
to match OVM.metadata.labels["kubevirt.io/domain"].

The RDP link is rendered if:
- matching Service is defined
- VM is running
- VM is Windows (OVM.metadata.labels["kubevirt.io/domain"] starts at "win")

If Service object can not be found, but the VM is Windows and running,
informative text about missing Service is rendered instead of the link.

The Service object exposing IP:Port should conform [1].
In a nutshell, the IP:Port is determined for Service type of:
- NodePort:
  - port = Service.spec.port[].nodePort
  - IP = nodeIP of node the VM is scheduled to (conforms pod.spec.hostIP)
- ClusterIP
  - port = Service.spec.port[].port
  - IP = Service.spec.clusterIP
- LoadBalancer
  - port = Service.spec.port[].port
  - IP = Service.spec.externalIP[0]

The Service is expected to map 3389 port to one of the above.

VM operating system is determined from "kubevirt.io/os" label of
the running OfflineVirtualMachine object.

[1] https://github.com/kubevirt/user-guide/blob/master/workloads/virtual-machines/expose-service.md